backend/user_service/requirements.txt
Flask
pymongo
flask-bcrypt
PyJWT
gunicorn
requests
flask-cors
